+++ src/trees_eval/A0_2Gsc_eval.cpp
@@ -566,7 +566,7 @@ template <int i0, int i1, int i2, class T> complex<T>  A2Gsc1g8_eval(const eval_
 
 
 
-template <class T> complex<T> (*A2Gsc1g_Tree_Ptr_eval(long hc))(const eval_param<T>&, const mass_param_coll&) {
+template <class T> complex<T> (*A2Gsc1g_Tree_Ptr_eval(long long hc))(const eval_param<T>&, const mass_param_coll&) {
 //    cout << "Found massive three-point vertex:" << hex << hc << dec << endl;
 	switch (hc) {// Note that the labels are in the reverse order of the respective process
 	case 0x413://G+g+G-
@@ -703,13 +703,13 @@ template <class T> complex<T> (*A2Gsc1g_Tree_Ptr_eval(long hc))(const eval_param
 
 
 
-template complex<R> (*A2Gsc1g_Tree_Ptr_eval(long hc))(const eval_param<R>&, const mass_param_coll&);
-template complex<RHP> (*A2Gsc1g_Tree_Ptr_eval(long hc))(const eval_param<RHP>&, const mass_param_coll&);
-template complex<RVHP> (*A2Gsc1g_Tree_Ptr_eval(long hc))(const eval_param<RVHP>&, const mass_param_coll&);
+template complex<R> (*A2Gsc1g_Tree_Ptr_eval(long long hc))(const eval_param<R>&, const mass_param_coll&);
+template complex<RHP> (*A2Gsc1g_Tree_Ptr_eval(long long hc))(const eval_param<RHP>&, const mass_param_coll&);
+template complex<RVHP> (*A2Gsc1g_Tree_Ptr_eval(long long hc))(const eval_param<RVHP>&, const mass_param_coll&);
 
 #if BH_USE_GMP
 
-template complex<RGMP> (*A2Gsc1g_Tree_Ptr_eval(long hc))(const eval_param<RGMP>&, const mass_param_coll&);
+template complex<RGMP> (*A2Gsc1g_Tree_Ptr_eval(long long hc))(const eval_param<RGMP>&, const mass_param_coll&);
 #endif
